// AccutracWebsite/src/app/core/interfaces/job-address/job-address-info.ts
export interface JobAddressInfo {
    companyName?: string;
    company_Code?: string;
    customer?: string;
    customerId?: number;
    email?: string;
    geoZoneId?: number;
    geoZoneName?: string;
    jobAddress?: string;
    jobAddressId?: number;
    jobCity?: string;
    jobState?: string;
    jobZip?: string;
    mailingAddress?: string;
    mailingCity?: string;
    mailingState?: string;
    mailingZip?: number;
    notes?: string;
    phone: string;
    phoneType?: number;
    zoneId?: number;
    zoneName?: string;
}